nrf_802154: rev 62eb9ffddf67f6d33fb6e2134227d60919e89888
This commit updates revision of the nrf_802154 component.
